Created generic controllers that take in command dictionaries
